+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v2.7.0
 - Added support for custom pelts
+- Added support for converting audio files to AudioClip objects
+- Added support for adding custom tracks to the Gramophone
 
 ## v2.6.0
 - Added support for custom consumable items using a choice of a few models